troca de float para double

+++ Exercício1/PontoCartesiano.java
@@ -1,26 +1,26 @@
 public class PontoCartesiano{
 
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public PontoCartesiano(float x, float y){
+    public PontoCartesiano(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public float getX(){
+    public double getX(){
         return x;
     }
 
-    public float getY(){
+    public double getY(){
         return y;
     }
 
-    public void setX(float x){
+    public void setX(double x){
         this.x = x;
     }
 
-    public void setY(float y){
+    public void setY(double y){
         this.y = y;
     }
